♻️ Fold collaboration back into general settings
The dedicated page held one switch with no credible growth path: poll
collaborators are managed on the poll, and the org layer replaces this
model rather than extending the page. A permanent nav entry is too
expensive for a set once toggle, and the support sentence (enable Team
collaboration in your space settings) maps to General anyway. The
section sits between Branding and the danger zone; discoverability at
the moment it matters is the invite flow follow up.
