+++ apps/docs/spaces/collaboration.mdx
@@ -21,8 +21,9 @@ This makes spaces useful in two different ways: as a shared workspace for a team
 ## Turning collaboration on or off
 
 <Steps>
-  <Step title="Open collaboration settings">
-    Go to **Settings → Collaboration**. You need to be an admin of the space.
+  <Step title="Open your space settings">
+    Go to **Settings → General** and find the **Collaboration** section. You
+    need to be an admin of the space.
   </Step>
   <Step title="Toggle Team collaboration">
     Turn the **Team collaboration** switch on or off, then press **Save**.
