+++ apps/web/src/app/[locale]/(space)/settings/components/sidebar.tsx
@@ -15,7 +15,6 @@ import {
   BoltIcon,
   CalendarIcon,
   CreditCardIcon,
-  HandshakeIcon,
   KeyIcon,
   LockIcon,
   PanelsTopLeftIcon,
@@ -121,12 +120,6 @@ export function SpaceSidebarMenu() {
       label: t("general", { defaultValue: "General" }),
       icon: <BoltIcon />,
       href: "/settings/general",
-    },
-    {
-      id: "collaboration",
-      label: t("collaboration", { defaultValue: "Collaboration" }),
-      icon: <HandshakeIcon />,
-      href: "/settings/collaboration",
     },
     // The members page is a management surface; members meet the roster
     // through in-context pickers instead.
